# go_file/src/ginDOme/docs/swagger.yaml
definitions:
  models.User:
    properties:
      account:
        description: Account 登录账号
        type: string
      password:
        description: Password 登录密码
        type: string
      re_password:
        description: RePassword 校验登录密码
        type: string
      state:
        description: State 用户状态， -1 - 异常；0 - 锁定；1 - 正常；
        type: integer
      user_profile:
        allOf:
        - $ref: '#/definitions/models.UserProfile'
        description: UserProfile 用户信息
    required:
    - account
    - password
    type: object
  models.UserProfile:
    properties:
      address:
        description: 地址
        type: string
      age:
        description: 年龄
        type: integer
      email:
        description: 邮箱
        type: string
      id_card:
        description: 身份证号
        type: string
      name:
        description: 账号名称
        type: string
      number:
        description: 手机号
        type: string
      sex:
        description: 性别
        type: integer
      user_id:
        description: User关联外键
        type: integer
    type: object
  pkg.ResCode:
    enum:
    - 1000
    - 1001
    - 1002
    - 1003
    - 1004
    - 1005
    - 1006
    - 1007
    - 1008
    - 1009
    type: integer
    x-enum-varnames:
    - CodeSuccess
    - CodeInvalidParam
    - CodeUserExist
    - CodeUserNotExist
    - CodeInvalidPassword
    - CodeServerBusy
    - CodeNeedLogin
    - CodeInvalidToken
    - CodeTokenCreation
    - CodeWrongCredentials
info:
  contact: {}
paths:
  /login:
    post:
      consumes:
      - application/json
      description: 登录账号
      parameters:
      - description: 账户
        in: body
        name: account
        required: true
        schema:
          type: string
      - description: 密码
        in: body
        name: password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses: {}
      summary: 登录账号
      tags:
      - Users
  /register:
    post:
      consumes:
      - application/json
      description: 注册账户
      parameters:
      - description: 账户
        in: body
        name: account
        required: true
        schema:
          type: string
      - description: 密码
        in: body
        name: password
        required: true
        schema:
          type: string
      - description: 确认密码
        in: body
        name: re_password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses: {}
      summary: 注册账户
      tags:
      - Users
  /user:
    get:
      consumes:
      - application/json
      description: 获取用户列表
      parameters:
      - description: 页码
        in: query
        name: page
        required: true
        type: integer
      - description: 每页数量
        in: query
        name: size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/pkg.ResCode'
      summary: 获取用户列表
      tags:
      - Users
  /user/{id}:
    get:
      consumes:
      - application/json
      description: 获取用户信息
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      summary: 获取用户信息
      tags:
      - Users
swagger: "2.0"